// src/scale.d.ts
export declare namespace ScaleType {
    const LINEAR: 'linear';
    const LOG: 'log';
    const POW: 'pow';
    const SQRT: 'sqrt';
    const TIME: 'time';
    const UTC: 'utc';
    const SEQUENTIAL: 'sequential';
    const QUANTILE: 'quantile';
    const QUANTIZE: 'quantize';
    const THRESHOLD: 'threshold';
    const ORDINAL: 'ordinal';
    const POINT: 'point';
    const BAND: 'band';
}
export declare type ScaleType = typeof ScaleType.LINEAR | typeof ScaleType.LOG | typeof ScaleType.POW | typeof ScaleType.SQRT | typeof ScaleType.TIME | typeof ScaleType.UTC | typeof ScaleType.SEQUENTIAL | typeof ScaleType.ORDINAL | typeof ScaleType.POINT | typeof ScaleType.BAND;
// ...
export declare type NiceTime = 'second' | 'minute' | 'hour' | 'day' | 'week' | 'month' | 'year';
export interface ScaleConfig {
    /**
     * If true, rounds numeric output values to integers.
     * This can be helpful for snapping to the pixel grid.
     * (Only available for `x`, `y`, `size`, `row`, and `column` scales.)
     */
    round?: boolean;
    /**
     * If true, values that exceed the data domain are clamped to either the minimum or maximum range value
     */
    clamp?: boolean;
    /**
     *  Default range step for `x` ordinal scale when is mark is `text`.
     *  @minimum 0
     */
    textXRangeStep?: number;
    /**
     * Default range step for (1) `y` ordinal scale,
     * and (2) `x` ordinal scale when the mark is not `text`.
     *
     * @minimum 0
     * @nullable
     */
    rangeStep?: number | null;
    /**
     * Default inner padding for `x` and `y` band-ordinal scales.
     * @minimum 0
     * @maximum 1
     */
    bandPaddingInner?: number;
    /**
     * Default outer padding for `x` and `y` band-ordinal scales.
     * If not specified, by default, band scale's paddingOuter is paddingInner/2.
     * @minimum 0
     * @maximum 1
     */
    bandPaddingOuter?: number;
    /**
     * Default outer padding for `x` and `y` point-ordinal scales.
     * @minimum 0
     * @maximum 1
     */
    pointPadding?: number;
    /**
     * Default spacing between faceted plots.
     * @TJS-type integer
     * @minimum 0
     */
    facetSpacing?: number;
    /**
     * Uses the source data range as scale domain instead of aggregated data for aggregate axis.
     * This property only works with aggregate functions that produce values within the raw data domain (`"mean"`, `"average"`, `"stdev"`, `"stdevp"`, `"median"`, `"q1"`, `"q3"`, `"min"`, `"max"`). For other aggregations that produce values outside of the raw data domain (e.g. `"count"`, `"sum"`), this property is ignored.
     */
    useRawDomain?: boolean;
}